+++ webservice/sources/__init__.py
@@ -1,0 +1,4 @@
+from .GHCC import GHCC
+from .JUST import JUST
+from .RVSCDE import RVSCDE
+from .OPENJUSTICE import OPENJUSTICE