+++ CHANGELOG.md
@@ -7,6 +7,11 @@ This project adheres to [Semantic Versioning](http://semver.org/)
 and to the [CHANGELOG recommendations](http://keepachangelog.com/).
 ## Alpha Development
 
+### [0.2.0] - (2020-11-20)
+- Merge ressources offered by different sources
+- Ordered output
+- Linked to OpenJustice document api
+
 ### [0.1.1] - (2020-11-11)
 - Added changelog
 - IUBEL link to pdf and extraction